Updates: Add boot partition updates to create-update-payload
The update payload creation script now supports creating a payload for
the boot partitions and BLFS on raspberry pi 3.

Rework CLI to use 'bootloader components' and 'kernel' instead of
'bootloader slots' and 'boot partition', so users don't need to worry
about the details of the location of the components to update on the target.

Users now just specify either bootloader component 1,2 or kernel.

JIRA: IOTMBL-2284/2405