meta-mbl-distro:recipes-support: Add swupdate package (#757)
We're using swupdate for robust update. This patch adds swupdate to
the packagegroup so it is included in the image.

We also need to configure swupdate so it doesn't depend on u-boot or
libubootenv as we don't have a u-boot environment. This patch also
adds a defconfig which sets CONFIG_BOOTLOADER_NONE and unsets
CONFIG_UBOOT so swupdate doesn't try to write to the
u-boot environment. We also remove some other configs which aren't required.